test: align iOS error taxonomy harness
