+++ VCL/VCL/api/entities/VCLJwt.swift
@@ -1,6 +1,6 @@
 //
 //  VCLJwt.swift
-//  
+//
 //
 //  Created by Michael Avoyan on 26/04/2021.
 //
@@ -15,9 +15,9 @@ public struct VCLJwt {
     public private(set) var payload: [String: Any]? = nil
     public private(set) var signature: String? = nil
     public private(set) var encodedJwt: String = ""
-    
+
     public private(set) var jwsToken: JwsToken<VCLClaims>? = nil
-    
+
     public init(
         header: [String: Any]?,
         payload: [String: Any]?,
@@ -26,7 +26,7 @@ public struct VCLJwt {
     ) {
         initialize(header: header, payload: payload, signature: signature, encodedJwt: encodedJwt)
     }
-    
+
     public init(encodedJwt: String) {
         let decodedJwt = encodedJwt.decodeJwtBase64Url()
         if decodedJwt.count == 3 {
@@ -42,7 +42,29 @@ public struct VCLJwt {
             )
         }
     }
-    
+
+    public init(validatedEncodedJwt encodedJwt: String) throws {
+        let jwtParts = encodedJwt.split(separator: ".", omittingEmptySubsequences: false).map { String($0) }
+        guard jwtParts.count == 3 else {
+            throw VCLError(message: "JWT must contain header, payload, and signature")
+        }
+        guard !jwtParts[0].isEmpty, let headerJson = jwtParts[0].decodeBase64URL(), let header = headerJson.toDictionary() else {
+            throw VCLError(message: "Failed to decode JWT header")
+        }
+        guard !jwtParts[1].isEmpty, let payloadJson = jwtParts[1].decodeBase64URL(), let payload = payloadJson.toDictionary() else {
+            throw VCLError(message: "Failed to decode JWT payload")
+        }
+        guard !jwtParts[2].isEmpty else {
+            throw VCLError(message: "JWT signature is missing")
+        }
+        initialize(
+            header: header,
+            payload: payload,
+            signature: jwtParts[2],
+            encodedJwt: encodedJwt
+        )
+    }
+
     private mutating func initialize(
         header: [String: Any]? = nil,
         payload: [String: Any]? = nil,
@@ -53,23 +75,23 @@ public struct VCLJwt {
         self.payload = payload
         self.signature = signature
         self.encodedJwt = encodedJwt
-        
+
         self.jwsToken = JwsToken<VCLClaims>(from: encodedJwt)
     }
-    
+
     public struct CodingKeys {
         public static let KeyTyp = "typ"
         public static let KeyAlg = "alg"
         public static let KeyKid = "kid"
         public static let KeyJwk = "jwk"
-        
+
         public static let KeyX = "x"
         public static let KeyY = "y"
-        
+
         public static let KeyHeader = "header"
         public static let KeyPayload = "payload"
         public static let KeySignature = "signature"
-        
+
         public static let KeyIss = "iss"
         public static let KeyAud = "aud"
         public static let KeySub = "sub"
@@ -79,7 +101,7 @@ public struct VCLJwt {
         public static let KeyExp = "exp"
         public static let KeyNonce = "nonce"
     }
-    
+
     var kid: String? { get {
         return (header?[CodingKeys.KeyKid] as? String) ?? ((header?[CodingKeys.KeyJwk] as? [String: Any])?[CodingKeys.KeyKid]) as? String
     } }
