+++ VCL/VCL/impl/data/usecases/CredentialManifestUseCaseImpl.swift
@@ -1,6 +1,6 @@
 //
 //  CredentialManifestUseCaseImpl.swift
-//  
+//
 //
 //  Created by Michael Avoyan on 09/05/2021.
 //
@@ -10,13 +10,13 @@
 import Foundation
 
 final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
-    
+
     private let credentialManifestRepository: CredentialManifestRepository
     private let resolveDidDocumentRepository: ResolveDidDocumentRepository
     private let jwtServiceRepository: JwtServiceRepository
     private let credentialManifestByDeepLinkVerifier: CredentialManifestByDeepLinkVerifier
     private let executor: Executor
-    
+
     init(
         _ credentialManifestRepository: CredentialManifestRepository,
         _ resolveDidDocumentRepository: ResolveDidDocumentRepository,
@@ -30,7 +30,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
         self.credentialManifestByDeepLinkVerifier = credentialManifestByDeepLinkVerifier
         self.executor = executor
     }
-    
+
     func getCredentialManifest(
         credentialManifestDescriptor: VCLCredentialManifestDescriptor,
         verifiedProfile: VCLVerifiedProfile,
@@ -43,37 +43,12 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
                 [weak self] credentialManifestResult in
                 guard let self = self else { return }
                 do {
-                    let credentialManifest = VCLCredentialManifest(
-                        jwt: VCLJwt(encodedJwt: try credentialManifestResult.get()),
-                        vendorOriginContext: credentialManifestDescriptor.vendorOriginContext,
+                    self.decodeCredentialManifestJwt(
+                        encodedJwt: try credentialManifestResult.get(),
+                        credentialManifestDescriptor: credentialManifestDescriptor,
                         verifiedProfile: verifiedProfile,
-                        deepLink: credentialManifestDescriptor.deepLink,
-                        didJwk: credentialManifestDescriptor.didJwk,
-                        remoteCryptoServicesToken: credentialManifestDescriptor.remoteCryptoServicesToken
+                        completionBlock: completionBlock
                     )
-                    resolveDidDocumentRepository.resolveDidDocument(did: credentialManifest.iss) { [weak self] didDocumentResult in
-                        do {
-                            let didDocument = try didDocumentResult.get()
-                            if let error = self?.validateDidDocument(didDocument, credentialManifest: credentialManifest) {
-                                self?.onError(error, completionBlock)
-                                return
-                            }
-                            self?.verifyCredentialManifestJwt(
-                                credentialManifest,
-                                didDocument,
-                                completionBlock
-                            )
-                        } catch {
-                            self?.onError(
-                                ErrorTaxonomy.classifyDidResolution(
-                                    VCLError(error: error),
-                                    requestKind: ErrorTaxonomy.requestKindIssuing,
-                                    requestDid: credentialManifest.iss
-                                ),
-                                completionBlock
-                            )
-                        }
-                    }
                 } catch {
                     self.onError(
                         ErrorTaxonomy.classifyClientRequestFetch(
@@ -87,7 +62,70 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
             }
         }
     }
-    
+
+    private func decodeCredentialManifestJwt(
+        encodedJwt: String,
+        credentialManifestDescriptor: VCLCredentialManifestDescriptor,
+        verifiedProfile: VCLVerifiedProfile,
+        completionBlock: @escaping (VCLResult<VCLCredentialManifest>) -> Void
+    ) {
+        jwtServiceRepository.decodeJwt(encodedJwt: encodedJwt) { [weak self] jwtResult in
+            guard let self = self else { return }
+            do {
+                let credentialManifest = VCLCredentialManifest(
+                    jwt: try jwtResult.get(),
+                    vendorOriginContext: credentialManifestDescriptor.vendorOriginContext,
+                    verifiedProfile: verifiedProfile,
+                    deepLink: credentialManifestDescriptor.deepLink,
+                    didJwk: credentialManifestDescriptor.didJwk,
+                    remoteCryptoServicesToken: credentialManifestDescriptor.remoteCryptoServicesToken
+                )
+                self.resolveCredentialManifestDid(
+                    credentialManifest,
+                    completionBlock
+                )
+            } catch {
+                self.onError(
+                    ErrorTaxonomy.classifyRequestValidation(
+                        VCLError(error: error),
+                        requestKind: ErrorTaxonomy.requestKindIssuing,
+                        requestDid: nil
+                    ),
+                    completionBlock
+                )
+            }
+        }
+    }
+
+    private func resolveCredentialManifestDid(
+        _ credentialManifest: VCLCredentialManifest,
+        _ completionBlock: @escaping (VCLResult<VCLCredentialManifest>) -> Void
+    ) {
+        resolveDidDocumentRepository.resolveDidDocument(did: credentialManifest.iss) { [weak self] didDocumentResult in
+            do {
+                let didDocument = try didDocumentResult.get()
+                if let error = self?.validateDidDocument(didDocument, credentialManifest: credentialManifest) {
+                    self?.onError(error, completionBlock)
+                    return
+                }
+                self?.verifyCredentialManifestJwt(
+                    credentialManifest,
+                    didDocument,
+                    completionBlock
+                )
+            } catch {
+                self?.onError(
+                    ErrorTaxonomy.classifyDidResolution(
+                        VCLError(error: error),
+                        requestKind: ErrorTaxonomy.requestKindIssuing,
+                        requestDid: credentialManifest.iss
+                    ),
+                    completionBlock
+                )
+            }
+        }
+    }
+
     private func verifyCredentialManifestJwt(
         _ credentialManifest: VCLCredentialManifest,
         _ didDocument: VCLDidDocument,
@@ -125,7 +163,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
             }
         }
     }
-    
+
     private func validateDidDocument(
         _ didDocument: VCLDidDocument,
         credentialManifest: VCLCredentialManifest
@@ -139,7 +177,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
         }
         return nil
     }
-    
+
     private func missingJwtKidError(requestDid: String?) -> VCLError {
         ErrorTaxonomy.classifyRequestValidation(
             VCLError(message: "JWT kid is missing"),
@@ -147,7 +185,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
             requestDid: requestDid
         )
     }
-    
+
     private func unresolvedJwtKeyError(kid: String, requestDid: String?) -> VCLError {
         ErrorTaxonomy.classifyRequestValidation(
             VCLError(message: "public jwk not found for kid: \(kid)"),
@@ -155,7 +193,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
             requestDid: requestDid
         )
     }
-    
+
     private func verifyCredentialManifestByDeepLink(
         _ credentialManifest: VCLCredentialManifest,
         _ didDocument: VCLDidDocument,
@@ -196,7 +234,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
             }
         }
     }
-    
+
     private func onVerificationSuccess(
         _ isVerified: Bool,
         _ credentialManifest: VCLCredentialManifest,
@@ -217,7 +255,7 @@ final class CredentialManifestUseCaseImpl: CredentialManifestUseCase {
             )
         }
     }
-    
+
     private func onError(
         _ error: Error,
         _ completionBlock: @escaping (VCLResult<VCLCredentialManifest>) -> Void
